WrapperGameClientInteractEntity created; Some debugs removed; compatInjector setting removed;

+++ src/main/java/io/github/retrooper/packetevents/wrapper/PacketWrapper.java
@@ -151,7 +151,7 @@ public class PacketWrapper {
         return byteBuf.readLong();
     }
 
-    public double readFloat() {
+    public float readFloat() {
         return byteBuf.readFloat();
     }
